Fixed build

+++ rviz_rendering/src/rviz_rendering/objects/screw_visual.cpp
@@ -102,7 +102,8 @@ void ScrewVisual::setScrew(const Ogre::Vector3 & linear, const Ogre::Vector3 & a
       Ogre::Vector3 point =
         Ogre::Vector3(
         static_cast<float>((angular_length / 4.0f) * cos(i * 2.0f * M_PI / 32.0f)),
-        static_cast<float>((angular_length / 4.0f) * sin(i * 2.0f * M_PI / 32.0f), angular_length / 2.0f));
+        static_cast<float>((angular_length / 4.0f) * sin(i * 2.0f * M_PI / 32.0f)),
+        static_cast<float>(angular_length / 2.0f));
       circle_angular_->addPoint(orientation * point);
     }
   }
